% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altlist.R
\name{rowwise}
\alias{rowwise}
\title{Iterate over the rows of a data frame}
\usage{
rowwise(x)
}
\arguments{
\item{x}{Data frame to iterate over.}
}
\value{
ALTREP object, that behaves like list of rows of a data frame.
  List elements also have an `.idx` entry, which is a scalar integer,
  the row number.
}
\description{
See examples below.
}
\examples{
for (el in rowwise(mtcars)) {
  print(paste0("row #", el$.idx))
}

lapply(rowwise(mtcars), function(elmt) { elmt$wt })
}
